requirements.txt
streamlit
numpy
opencv-python==4.8.1.78
mahotas
scikit-image
Pillow
lightgbm

